Wand class (#11)
* Added base spell class outline

* Added spell stat struct and list

* Created projectile abstract class outline

* Implemented projectile functions, added test implementations

* Added test for casting multiple spells and direct Spell inheritance

* Added wand basics with test scene

* Updated test scene

* Added mana consideration

+++ Assets/Scripts/Spellcasting/SpellStatsSO.cs
@@ -7,7 +7,7 @@ public struct SpellStat
     public string spellName;
     public float damage;
     public float cooldown;
-    public float mana;
+    public int mana;
     public float speed;
     public float range;
     public float area;
